Impute Lambda role ARN rather than looking it up
The lookup will fail when the role hasn’t been created yet, and there’s no way to specify that the passed-in role is a dependency of the module, creating a kind of bootstrapping paradox.

This is less-elegant but should work. It may still run into consistency issues but those will be resolved with a re-run.